Old virtual list code removed
