+++ src/ui/ListOfAssets/Multi.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { Asset } from './api'
 
-  import VirtualList from '@/ui/VirtualList/_index.svelte'
+  import VirtualList from '@/ui/VirtualList/index.svelte'
   import Layout from './Layout.svelte'
   import { TABS } from './Tabs.svelte'
   import AssetItem from './CheckboxAsset.svelte'
